Primitive automated CMS thing
Unfortunately, this is not very useful but it's a bit more
convenient.

We cannot edit any arbitrary article and not support previewing what
we render (I commit on any finish), which is fine if I write perfectly.

I can technically still separate the steps... Which is fine, however I'd
prefer not to do that.

For now this works fine, since I write only text. I may just use a type
of markdown to generate my pages I suppose.

The script is written in bash and generates text files in the correct
format hopefully.