Fixed page linkage
I forgot to account for the fact that my page-links system works by
taking an assumed sorted list. So obviously the sorting had to be pushed
out.

Unfortunately this makes the files get loaded twice. However this should
lead to a nice cleanup of the system quite naturally due to this new
scheme.